More testing and fuzzing and docs for pools (#12624)
* move pools fuzzing to hongfuzz

* merge more small fixes

* fix all tests

* Update frame/nomination-pools/fuzzer/src/call.rs


* remove transactional

* fmt

* fix CI

* fmt

* fix build again

* fix CI
